Add launch.py module and update run.py with configuration support
